Switch from FzfLua to Telescope
Telescope seems to have a nicer UI, and I like that the file previews
are native buffers as opposed to relying on bat. I was convinced from
this video showcasing the differences:

    https://www.youtube.com/watch?v=8sba2Veu_uE

To be seen if there is a speed difference.

Also trying out Oil.nvim, which seems great. Though it seems to break
netrw use unexplicably.

Cleanup tmux config a bit. "bind" is just an alias for "bind-key", be
more consistent.